+++ packages/affiliation-merchant-addresses/src/component/MerchantAddresses.js
@@ -44,7 +44,7 @@ export const AffiliationMerchantAddresses = (Base = class {}) => class extends
     if (affiliationCode) {
       this
         .request([
-          sdk.merchants.addresses.get({ affiliationCode }),
+          sdk.affiliation.addresses.get({ affiliationCode }),
         ])
         .then((responses) => {
           if (responses.every(notEmpty)) {