+++ packages/affiliation-merchant-addresses-list/CHANGELOG.md
@@ -1,3 +1,11 @@
+# [affiliation-merchant-addresses-list-v1.1.0](https://github.com/stone-payments/affiliation-web-components/compare/affiliation-merchant-addresses-list-v1.0.2...affiliation-merchant-addresses-list-v1.1.0) (2019-02-04)
+
+
+### Features
+
+* :construction: Update sdk method. ([f328e79](https://github.com/stone-payments/affiliation-web-components/commit/f328e79))
+* :construction: Update view. ([fe299bb](https://github.com/stone-payments/affiliation-web-components/commit/fe299bb))
+
 # [affiliation-merchant-addresses-list-v1.0.2](https://github.com/stone-payments/affiliation-web-components/compare/affiliation-merchant-addresses-list-v1.0.1...affiliation-merchant-addresses-list-v1.0.2) (2018-11-19)